requirements.txt
moviepy>=2.0.0
Pillow>=10.0.0
typer>=0.9.0
pydantic>=2.0.0
pyyaml>=6.0.0
openai>=1.0.0
httpx>=0.25.0
rich>=13.0.0
streamlit>=1.30.0
